Revert "fix(daemon): relocate daemon host image out of the install-dir kill zone (#7473)"
This reverts commit f4faafa9874c188be461d3138c9040b6332b44c1.
